// PropertyRenting.Api/ClientApp/src/app/Models/voucher-details.ts
export interface VoucherDetails {
    id?: any;
    accountId?: any;
    debitAmount?: number;
    creditAmount?: number;
    renterId?: any;
    ownerId?: any;
    contributerId?: any;
    expenseId?: any;
    buildingId?: any;
    unitId?: any;
    cashBankId?: any;
    account?: string;
    renter?: string;
    owner?: string;
    contributer?: string;
    building?: string;
    unit?: string;
    cashBank?: string;
    expense?: string;
}